Site: Improve FIT files parsing, display device and firmware version in activity view

+++ database/migrations/2023_05_23_022428_create_device_manufacturers.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('device_manufacturers', function (Blueprint $table) {
-            $table->integer('id');
+            $table->bigInteger('id');
             $table->string('manufacturer');
             $table->string('description');
             $table->timestamps();
